cinearenas/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
django-phonenumber-field==7.3.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
gunicorn==21.2.0
mercadopago==2.2.1
qrcode==7.4.2
pypng==0.20220715.0
pillow==10.2.0
django-jsonfield==1.4.1
